requirements.txt
Django==3.2
django-cors-headers==3.7.0
djangorestframework==3.12.4
matplotlib==3.3.4
numpy==1.19.5
opencv-python==3.4.9.31
Pillow==8.2.0
torch==1.8.1+cpu
torchaudio==0.8.1
torchsummary==1.5.1
torchvision==0.9.1+cpu
